requirements.txt
numpy==2.4.2
pandas==3.0.1
pdf2image==1.17.0
pillow==12.1.1
PyMuPDF==1.27.1
python-dateutil==2.9.0.post0
six==1.17.0
